Add files via upload
if else statements based on bool values to display class times in an alert window. still working on making the buttons sit side by side. also the word this is a placeholder for values if we need to fancy it up more and make sure our functions are working correctly